Don't tie error display to invalid state (#407)
## 🤨 Rationale

Currently our CSS for the text field displays the error style, icon, and error text when either the `invalid` class has been applied or when the `:invalid` pseudo-class applies. The latter is triggered by form validation, which sometimes occurs sooner than we want to display the error styling. For example, a required input control should not immediately display an error; it should give the user a chance to enter a value first. It should only display the error after the user has focused and then left the control, or after attempting to submit the form. To give users better control over when validation errors are displayed, we should not trigger that styling on the `:invalid` pseudo-class.

## 👩‍💻 Implementation

Removed CSS selectors for the `:invalid` pseudo-class.

## 🧪 Testing

Manual testing in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.