Icon button use hidden content as a11y name (#438)
## 🤨 Rationale

Fixes #333 

## 👩‍💻 Implementation

The previous attempt to make this work referenced the hidden `span.content` from the `nimble-button`'s `aria-labelledby` attribute. However, we were setting the `span.content`'s `display` to `none`. This removes it from the rendered tree, preventing accessibility tools from reading the label.

I initially looked into ways of finding the (hidden) button label and explicitly setting `aria-label` on the `nimble-button`, but that was fragile and ugly.

Instead, I found an article that recommended making the element effectively invisible without actually removing it from the rendering tree. This allows the button to use that as the a11y name, even without using `aria-labelledby`. So I removed that logic, too.

## 🧪 Testing

Tested in Storybook with the dev tools a11y tree and tab. Tested that NVDA pronounced the name.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.