Fix button control's width & height (#453)
# Pull Request

## 🤨 Rationale

With the existing button styles, setting a min-width or a min-height on a `nimble-button` does not behave as expected. The width/height of the `nimble-button` itself grows to match the minimums specified in the style, but the inner control's width and height do not grow to match. The result is that the clickable region of the button is a different size than the styled region of the button.

## 👩‍💻 Implementation

Setting the width and height of the inner control to 100% rather than 'inherit' ensures that the size of the inner control will match the size of the button itself.

## 🧪 Testing

Used dev tools to set a min height and min width on the `nimble-button` in storybook and ensured that the styled portion of the button grew appropriately.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.