Update fast dependencies (#419)
# Pull Request

## 🤨 Rationale

There are some changes we've been waiting on from FAST. Notably a [re-work of the FAST tree-view events](https://github.com/microsoft/fast/pull/5350). Additionally, the newest version of `fast-foundation` allows us to remove all of the `@ts-expect-error` compiler flags that we had to insert to avoid build errors.

## 👩‍💻 Implementation

Just removed the compiler flags, and updated the Nimble `TreeItem` to handle the updated re-working of the FAST `TreeView`.

## 🧪 Testing

Ran current tests.