nimble-text-area should allow setting width (#457)
## 🤨 Rationale

The `nimble-text-area` cannot be made to stretch to fill the width of its container. Or rather, the contained `textarea` does not obey the width set on the `nimble-text-area`.

## 👩‍💻 Implementation

Setting `width` to `inherit` on the `textarea` (class `control`) causes it to inherit any `width` value set on the `nimble-text-area` element. Note that `nimble-text-area` also supports controlling the visible width via a `cols` attribute (number of visible text columns). When the `width` style is set, this overrides the `cols` attribute.

## 🧪 Testing

Tested manually in dev tools.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.