Add ability to place buttons within a text field (#467)
# Pull Request

## 🤨 Rationale

We should allow buttons to be slotted at the end of a text field. Brandon's designs have a "clear" button on some text fields, and we have also already come up with a use-case for an "edit" button.
#463

## 👩‍💻 Implementation

The styling for the text field assumed the only content within the `end` slot would be the error icon & text, so I put an `error-content` class on the error icon & text and updated the styling accordingly. Additionally, I added a new slot within the `end` slot that clients can add buttons to.

## 🧪 Testing
- Added button control to the text-field story
- Added buttons to the text-field matrix.
  - Note: I put the buttons in the matrix story as 'outline' even though I expect most usages to be 'ghost'. The purpose of this is to be able to test that the dimensions of the button is expected.

## ✅ Checklist

- [ ] I have updated the project documentation to reflect my changes or determined no changes are needed.