Add npm audit step to main pipeline workflow (#475)
# Pull Request

## 🤨 Rationale

It seems wise to use the pipeline to enforce that our dependencies don't contain known security vulnerabilities. The [npm audit](https://docs.npmjs.com/cli/v8/commands/npm-audit) command is designed to report and enforce this.

The current state of the repo contains: 
 - 0 vulnerabilities in our production dependencies
 - several vulnerabilities in our `devDependencies` (including 1 critical) which could be resolved with minor version changes by running `npm audit fix`, but some of those changes involve downgrades to packages we probably don't want to downgrade.
 - several vulnerabilities in our `devDependencies` which will require major version bumps to fix (via `npm audit fix --force`)

[See this build](https://github.com/ni/nimble/runs/5794505245?check_suite_focus=true) for a report of the vulnerabilities in our devDependencies. You can use `npm ls package-name` locally to see what tree brings in a particular sub-dependency. Most of the ones I checked are from storybook or Angular 12.

## 👩‍💻 Implementation

In the interest of making high-value, low-cost progress, this PR only enables auditing of production dependencies in our PR/release workflow. (via `npm audit --only=prod`). As a follow-on step, we could start fixing the dev vulnerabilities (for example by upgrading to Angular 13).


## 🧪 Testing

Ran a build that verified the command will fail the pipeline if it detects any vulnerabilities.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.