Add styling for indeterminate checkbox state (#483)
## 🤨 Rationale

The Nimble checkbox is missing styling for the indeterminate (i.e. partial check) state. 

## 👩‍💻 Implementation

The underlying FAST checkbox does have an `indeterminate` property and setting it adds the `indeterminate` class to the element. It also provides an `indeterminate-indicator` slot for the icon to show. Note that FAST does not provide an attribute on the element, only the property on the view model.

In this change, we:
1. Add a "minus" icon to the `indeterminate-indicator` slot. There is no design to reference for what the icon should look like. Some prior art uses a square, and some use a minus. We do not currently have a square icon in Nimble.
2. Update the CSS to show the minus icon when the `indeterminate` class is present. This class trumps the checked state.

## 🧪 Testing

Added indeterminate state to checkbox matrix stories.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.