Fix intermittent  tree styling issues (#493)
# Pull Request

## 🤨 Rationale

We've seen intermittent tree styling issues (see [azdo bug 1919045](https://dev.azure.com/ni/DevCentral/_workitems/edit/1919045) and #482). From what I can tell, there is a race condition in the FAST tree where sometimes the `nested` class does not get applied to the tree items. As a result, some of our styling does not get applied to the tree items. However, it appears that the `nested` class is an indication that the tree items are nested within a tree view. Since we currently only support tree items within a tree view and don't have plans to support them outside of a tree view, I've moved all of the `nested` specific styling to always be applied.

## 👩‍💻 Implementation

Move `.nested` styling to more general selectors that will be applied whether or not the `nested` class is added to the tree item.

## 🧪 Testing

Looked at the tree in storybook and in SLE to ensure the styling is correct.

## ✅ Checklist

- [ ] I have updated the project documentation to reflect my changes or determined no changes are needed.