Add frameless appearance for text field (#509)
# Pull Request

## 🤨 Rationale

After talking to Brandon, we determined that we need a new text field appearance for situations like the `sl-details-panel`:
![image](https://user-images.githubusercontent.com/7282195/164331190-a0469962-b5e0-475f-a7cf-592ad0442997.png)

## 👩‍💻 Implementation

While we only need this new "frameless" appearance for the readonly mode of the text field, it simplifies the API to add it as another value of the existing `appearance` attribute (i.e. as opposed to introducing a new `readonly-appearance`). Now instead of all appearances having the same styling when in `readonly` mode, each one has a unique `readonly` styling. The one that used to be applied to all is now just the `outline` appearance's readonly style.

## 🧪 Testing

Tested in Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.