Support clear-inline-padding on text field to align label and value (#532)
# Pull Request

## 🤨 Rationale

The sl-details-panel needs to display text fields with the start of the value and label text aligned:
![image](https://user-images.githubusercontent.com/7282195/166556312-21023a46-fce2-4cae-969c-f342072b9bc0.png)


## 👩‍💻 Implementation

Introduced a new class `clear-inline-padding` that can be set on the nimble-text-field to remove any space between the left edge of the control and the start of the content (i.e. the value/placeholder, or the left icon, if present). It also removes any right padding.

My approach to styling is to:
1. introduce standard horizontal `gap` between all elements in flex box
2. introduce 0-width elements `::before` and `::after` the `root` div. This applies the gap spacing at the beginning and end.
4. hide the 0-width elements when the `clear-inline-padding` class is on the host element

## 🧪 Testing

Added to Storybook. Created a new matrix story that combines the clear-inline-padding state with a few other interesting attributes.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.