Regenerate package-lock.json via npm install (#547)
# Pull Request

## 🤨 Rationale

After #542 went in, running `npm install` on a clean clone of `main` results in a dirty package-lock.json. I'm not sure what happened, but possibly it was an incorrect merge.

## 👩‍💻 Implementation

Run `npm install` after pulling the latest and doing `git clean -fdx`. The changes look more aligned with `package.json` and nearby dependencies in `package-lock.json`.

## 🧪 Testing

N/A. Relying on pipeline.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.