Fix small cosmetic issues with tab activation/hover (#538)
## 🤨 Rationale

Fixes #478 

## 👩‍💻 Implementation

1. Made hover bg color of the active tab green instead of gray
2. Made focus and active indicators (horizontal lines) contained within 32px height of tab
3. Replaced animation when switching active tab. Now grows/shrinks from/to center of tab.

To do implement 2 and 3, I first disabled the FAST active tab indicator/animation by setting `activeindicator` to false in the constructor. Then I used the `::before` and `::after` pseudo elements to add the focus and active indicator lines. They had to be new, distinct elements (as opposed to borders/box shadows) because (1) their widths during animation do not coincide with the widths of the label slot or the host element, and (2) they must appear _within_ and _on top_ of the background of the host.

## 🧪 Testing

Tested in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.