Fix slide-out buttons not pinned to bottom on Safari (#549)
## 🤨 Rationale

Fixes AzDO bug: https://ni.visualstudio.com/DevCentral/_workitems/edit/1926875

## 👩‍💻 Implementation

The issue comes from how we were doing layout in the nimble drawer. Safari doesn't support the values we were using (e.g. "auto") for `grid-template-rows`, so the slotted `<section>` element didn't grow to take up most of the vertical space. This left the `<footer>` (with the cancel and OK buttons) immediately following the `<section>` rather than appearing at the bottom of the drawer.

My solution is to change from a grid layout to a flexbox. This avoids the issue on Safari, though it does have the drawback of being a breaking change if the client doesn't supply the `<header>`, `<section>`, and `<footer>` in that order. The grid layout would map the slotted elements to the appropriate grid row, but the flexbox layout will put them in the order given. Still, since the natural inclination is to give them in the order expected in the final layout, this is unlikely to actually break any client.

## 🧪 Testing

Tested locally in Chrome in the Systems Management app. I am unable to test Safari.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.