Focus menu button in capture phase of menu change event (#569)
# Pull Request

## 🤨 Rationale

By focusing the menu button in the capture phase of the menu 'change' event, the button will be focused by the time a client's 'change' handler for the menu is called. As a result, the client can choose to focus something on the page during their 'change' handler without that focus being immediately modified by the menu button.

## 👩‍💻 Implementation

Add _capture: true_ when adding the event handler so that the handler is executed during the capture phase of the event rather than the bubbling phase. I could not accomplish this in the template, so I had to put the logic for adding/removing the event handler in the MenuButton class.

## 🧪 Testing

- Added new unit test that I verified failed without my change

## ✅ Checklist

- [ ] I have updated the project documentation to reflect my changes or determined no changes are needed.