Import latest DSP (#583)
# Pull Request

## 🤨 Rationale

In trying to clean-up the nimble-tokens package during testing I found that recent (good) changes to npm 8 now cause the DSP extension to fail to find `style-dictionary`. I reported the behavior here: https://github.com/AdobeXD/design-system-package-dsp/issues/29#issuecomment-1138848269

This PR adds a workaround for that issue by adding `style-dictionary` as a pinned devDependency to the root package and with a different version pinned as a devDependency in the `nimble-tokens` package. This forces a copy to be installed in `node_modules` for `nimble-tokens` and an unused copy to be installed at `node_modules` in the monorepo root.

In addition, the `hex-rgb` version was pinned for `style-dictionary` as the latest `hex-rgb` version switches to es6 modules and breaks the `nimble-extensions.js` script.

Also removed the `assets` folder from the source tree that is created by the DSP. We currently are not leveraging the component definitions in the DSP, only the tokens, and this improves maintainability for now.

## 👩‍💻 Implementation

See above.

## 🧪 Testing

Built the DSP locally.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.