Include change event in Storybook for nimble-tabs (#585)
# Pull Request

## 🤨 Rationale

A client asked how to be notified when the user changes the active tab. I noticed that FAST supports this by firing the `change` event but Nimble doesn't document it.

## 👩‍💻 Implementation

Add `change` event to Actions tab of nimble-tabs story, similar to other components. 

## 🧪 Testing

Locally verified that the event is tracked in the Actions tab in Storybook when you click or keyboard navigate to a different nimble-tab.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.