Fix select drop-down clipping (#588)
## 🤨 Rationale
Fixes #464

## 👩‍💻 Implementation

The FAST select has  a `maxHeight` property that is calculated based on the space available. Previously, the select also defined a `maxHeightChanged()` method that would write the `maxHeight` value into a CSS variable, `--max-height`. We were then using this variable in our Nimble select's stylesheet to control the height of the drop-down. At some point, the FAST select removed the `maxHeightChanged()` method, so `--max-height` no longer got set. My change reintroduces this function, but in our Nimble extension of FAST's select.

## 🧪 Testing

Tested manually in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.
* Avoid magic numbers