Text/number fields' hover indicator should grow from center (#597)
## 🤨 Rationale

The text and number fields have a hover indicator of a thicker underline. The design doc dictates that this line should grow from the middle of the control outward. We already do this for the tab and the select.

Note that the design doc also specifies keyboard focus styling (a second bottom border) for text and number fields, but these controls don't have a distinct keyboard focus state. It's the same as clicking in the control. The number field was actually missing the colored underline when focused, so I added that, too.

## 👩‍💻 Implementation

Used the same approach as for the tab and select: a pseudo element (::after) with no content but a bottom border provides the hover indicator. We center it within the control, then animate its width changing from 0 to 100%.

I also tweaked the approach we take for styling the text field. Instead of having a bottom border by default and letting the block and frameless appearances remove it, I start with no border, and let each appearance add any borders necessary. This is a little easier to understand.

## 🧪 Testing

Tested in Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.