Fix disappearing menu button (#602)
# Pull Request

## 🤨 Rationale

In some cases, menu-buttons were incorrectly becoming content-hidden menu-buttons when they were hidden & re-shown in Angular apps. See [this SLE bug](https://dev.azure.com/ni/DevCentral/_workitems/edit/2013239).

While I'm not 100% sure what Angular is doing to aggravate the nimble bug, there is an issue in the nimble-menu-button template where we are setting `content-hidden` to `true`/`false` rather than dynamically apply the `content-hidden` attribute. This breaks the styling because our CSS is only looking for the presence of the `content-hidden` attribute, not a value associated with it (which is the same strategy nimble takes for other attributes, such as `disabled`).

## 👩‍💻 Implementation

Update the nimble-menu-button template to conditionally add `content-hidden` on the inner nimble-toggle-button rather than setting the value to `true`/`false`.

## 🧪 Testing

- Pulled a local build of nimble-components into the Systems Management app and verified it fixes the problem described in the bug.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.