Update peerDependency from beachball-lock-update to beachball (#600)
# Pull Request

## 🤨 Rationale

I noticed a critical vulnerability being reported when running `npm i` in Skyline `Web/Packages`. It's the same vulnerability that we fixed in #589 but coming from `beachball-lock-update` instead of `beachball` directly. We missed updating `beachball-lock-update` in that PR because its dependency was a `peerDependency` so vulnerabilities aren't reported by `npm audit`.

## 👩‍💻 Implementation

Update `beachball` version to same used by the rest of the repo and run `npm install`.

## 🧪 Testing

Relying on pipeline.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.