Combobox control value accessor (#630)
An initial attempt to integrate the Nimble Combobox in SLE ran into the problem of not having a proper Control Value Accessor in place for it. The `DefaultValueAccessor` was not sufficient to support the use case of the Combobox options being tied to model types (instead of just strings). This led to the Combobox not displaying anything for a selected value.

Additionally, the [`SelectControlValueAccessor`](https://github.com/angular/angular/blob/7e03fc90d61602d4383d68af66724827172b321e/packages/forms/src/directives/select_control_value_accessor.ts#L94) can not be leveraged because of the mechanism for turning a model value into a display string.