Tooltip states implementation (#637)
# Pull Request

## 🤨 Rationale
Creates additional states and icon states of the Nimble Tooltip. This makes understanding the purpose of a tooltip easier for users and clients.

## 👩‍💻 Implementation
As seen in previous commits, the new states of the tooltip were implemented in various ways. The chosen method of implementation was discussed with the team and regarded as the simplest and most efficient process. See #622 for more information.

## 🧪 Testing
For each theme, matrix testing was created for every state and icon state along with a short-text and long-text version of the tooltip.

## ✅ Checklist

- [X] I have updated the project documentation to reflect my changes or determined no changes are needed.