Combobox control value accessor (#660)
# Pull Request

## 🤨 Rationale

- #659

## 👩‍💻 Implementation

Essentially just having the `NimbleComboboxListOptionDirective` update the containing `NimbleCombobox` display value when it gets initialized. This is where the map gets populated in the `NimbleComboboxControlValueAccessor` that allows it to resolve a model value to a text display value.

## 🧪 Testing

Added a mirror set of tests for Reactive-based Angular forms. Verified that without the implementation changes, tests regarding setting initial display values failed, and passed with the changes in place.