Remove tab focus from button in combobox. (#669)
## 🤨 Rationale

I noticed recently that the `toggle-button` in the template of the `nimble-combobox` traps the forward tab focus (see the [Storybook page](https://ni.github.io/nimble/storybook/?path=/story/combobox--combobox)). Other components that contain buttons in their template (such as the `number-field`) also do not allow the buttons to receive tab focus. We lose not functionality from the combobox by disabling the tab focus, as the user does not need to interact with that button to display or hide the dropdown (this is accomplished through the arrow keys and <Esc>).