Combobox text input updates value. (#668)
# Pull Request

## 🤨 Rationale

I recently discovered while trying to integrate the `NimbleCombobox` into SLE that the form didn't re-validate while typing into the `Combobox`. The reason was fairly obvious, as the FAST `Combobox` currently does not update its value on text input (as the native `input` does as well as the FAST `TextField`). This was problematic because having the form re-validate on text input is pretty helpful. 

## 👩‍💻 Implementation

The root issue is solved by simply having `nimble-combobox` now update the `value` on `input`. However, changes were still needed on the `NimbleControlValueAccessorDirective` so that it responded to `input` events for the `onChange` registered handlers.

** Blazor **

Blazor currently has issues with handling events causing the DOM to re-render (see [here](https://github.com/dotnet/aspnetcore/issues/31549). For now, we will not bother with updating Blazor to handle similar scenarios.

## 🧪 Testing

Created unit tests.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or determined no changes are needed.