Change backdrop overlay color for dialog and drawer (#700)
# Pull Request

## 🤨 Rationale

Fixes #690 
Also updates overlay color for Drawer.

## 👩‍💻 Implementation

This change updates the backdrop overlay color for the dialog and drawer. Prior to my change the dialog was using the default (for the native dialog element), which was something like 10% black for all themes, and the Drawer was using popup border color, which was a dark color for light theme, but a light color for the dark/color themes. I talked to Brandon about what the color should be, and we settled on a dark color for all themes: Black at 30% and 60% opacity for the light and dark/color themes, respectively.

I introduced a new token, `modalBackdropColor`, to represent these colors. I also had to introduce a new helper function that would allow different themes to use different opacities.

## 🧪 Testing

Examined in Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or determined no changes are needed.