Combobox/Select doc update (#715)
## 🤨 Rationale

The aim of this PR is to address the issue outlined in the linked issue
for providing the proper guidance to a client for when to use a `Select`
versus a `Combobox`. While I didn't make specific documentation along
the lines of "use this when...", I've tried to provide more detail about
the individual component behaviors to help a client make a more informed
choice.