Documentation updates for open source (#749)
# Pull Request

## 🤨 Rationale

As part of #740 we want to improve our docs around where to have
discussions and how to contribute.

## 👩‍💻 Implementation

1. I'm adding a Discussions issue template to give non-NI people a place
to reach out since they can't access Teams. Updated README to link to
it.
2. Added a note to CONTRIBUTING explaining current situation with
fork-pull workflows
3. Removed some links that are NI-only and not very relevant any more.

## 🧪 Testing

I'll create a test discussion once this completes to make sure the issue
template is working.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.