More font assets (#760)
# Pull Request

## 🤨 Rationale

When working on the new `nimble-dialog` styling, I realized that our
"light" fonts did not render correctly. There was no difference between
specifying a font-weight of 300 (light) and a font-weight of 400
(normal). To fix the issue, I'm pulling in the light version of Source
Sans Pro and the semi-bold version of Source Sans Pro. The light version
allows us to have the correct fonts for subtitles, and the semi-bold
version addresses some rendering inconsistencies we've seen between
browsers for bold text.

## 👩‍💻 Implementation

- Pulled in SourceSansPro-Light.woff2 and SourceSansPro-SemiBold.woff2
- Updated `fonts.css` to point to the new fonts based on font-weight

## 🧪 Testing

- Verified my pending dialog changes look correct
- Looked at Chromatic diffs. There are quite a few, but they are all
minor font changes. This affects a large number of our components
because our component labels are semi-bold.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.