Fix drawer issues by rewriting nimble-drawer to use HTML dialog (#728)
# Pull Request

## 🤨 Rationale

Resolves #592 
Resolves #636 
Resolves #234 
Resolves #190 
Resolves #186 
Resolves #189

The `nimble-drawer` had a number of issues associated with it (see list
above) that can be resolved by moving away from the FAST dialog in favor
of the HTML `<dialog>` element. See spec here: #692

## 👩‍💻 Implementation

- Created custom template for the `nimble-drawer` that follows many of
the patterns established by the `nimble-dialog`
- Updated Blazor & Angular wrappers around the `nimble-drawer` element
- Updated Blazor and Angular example apps to:
- have their headers remain fixed at the top of the window when
scrolling the rest of the content on the page
- [Angular only] use the `nimble-menu-button` in the header instead of a
button + menu
- no longer show the drawer as a pinnable navigation pane since that
isn't the ended use case of the drawer
- Move `USER_DISMISSED` and `UserDismissed` to the `patterns` directory
so they can be shared by the drawer & the dialog. This avoids issues
trying to have duplicated exports from Angular.
- Made a static token for `largeDelay` to be used for the animation of
the dialog's background

## 🧪 Testing

- Created unit tests for the web component
- Updated Blazor & Angular tests
- Updated storybook/matrix tests
- Note: the matrix tests are currently broken in Firefox because
Chromatic isn't using a recent enough version of Firefox to support the
`<dialog>` element
- Manually verified in storybook
- Manually verified in a subset of SLE apps & in systemlink-lib-angular

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>