Temporarily disable npm audit of dev dependencies (#802)
# Pull Request

## 🤨 Rationale

As described in #801 we have a vulnerability that is causing `npm audit`
to fail the PR build, preventing submissions.

## 👩‍💻 Implementation

Comment out the audit step that's failing. We'll use the bug linked
above to track re-enabling it.

## 🧪 Testing

PR build

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.