Angular 13 (#708)
# Pull Request

## 🤨 Rationale

This is required to upgrade applications to 14 from 12 to avoid an issue
with npm involving conditional versioning, e.g. `^13 || ^14`. Clients
can depend on this version of nimble-angular when upgrading to 13 and
then on the next version supporting 14.

An upgrade guide will be published detailing the steps taken to perform
the upgrade. It will recommend upgrading from 12 to 14 in lock-step but
all at once to address builds and tests only once, with version 14.
Therefore, usage of this version of nimble-angular is expected to be
minimal.

See #684 for more information.

## 👩‍💻 Implementation

- Upgrade to Angular 13.
- Enumerate the [update
list](https://update.angular.io/?l=3&v=12.0-13.0).
- Compile libraries with Ivy partial compilation.

## Remaining Work Items

2. Publish a new version of the styleguide
[PR#82](https://github.com/ni/javascript-styleguide/pull/82).

## 🧪 Testing

- Verified that the linter passes
- Verified that the build passes
- Verified that the automated tests pass
- Verified that the example application works correctly

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.